fect: 한국-외국 분리 시작
