+++ pipe/korea/korea_websocket_client.py
@@ -2,9 +2,8 @@ import json
 import websockets
 
 from pipe.korea.korea_rest_client import KoreaExchangeRestAPI
-from common.exception import SocketRetryOnFailure
-from common.core.types import SubScribeFormat, ExchangeResponseData
-from common.client.websocket_interface import (
+from common.core.types import ExchangeResponseData
+from common.client.market_socket.websocket_interface import (
     WebsocketConnectionManager,
     MessageDataPreprocessing,
 )
@@ -61,27 +60,5 @@ class KoreaWebsocketConnection(WebsocketConnectionManager):
             target="korea",
             folder="korea",
             process=KoreaMessageDataPreprocessing(),
-        )
-
-    async def websocket_to_json(
-        self, uri: str, subs_fmt: SubScribeFormat, symbol: str, socket_type: str
-    ) -> None:
-        """말단 소켓 시작 지점"""
-
-        @SocketRetryOnFailure(
-            retries=3,
-            base_delay=2,
             rest_client=KoreaExchangeRestAPI(),
-            symbol=symbol,
-            uri=uri,
-            subs=subs_fmt,
         )
-        async def connection():
-            async with websockets.connect(
-                uri, ping_interval=30.0, ping_timeout=60.0
-            ) as websocket:
-                await self.socket_param_send(websocket, subs_fmt)
-                await self.handle_connection(websocket, uri)
-                await self.handle_message(websocket, uri, symbol, socket_type)
-
-        await connection()
